swap ERROR_NOINTERFACE for E_NOINTERFACE
According to my copy of winerror.h ERROR_NOINTERFACE and
ERROR_NO_PROC_FOUND should both be defined. The latta compiles
fine but not the former. E_NOINTERFACE is supposed to be
defined in the same file too so maybe that will compile, if not
I'll have to search for another suitable code