Merge pali's changes and swap out name of error 0
Went back to find the changes just a few minutes ago and the
link I found in my PR was this one:

https://github.com/dlfcn-win32/dlfcn-win32/issues/115#issuecomment-2481400024

Amounted to just adding spacing, though I coulda sworn previously
there was an intent to use LocalAlloc there instead of heaps.
Heaps are faster and more thread friendly though so I'll leave
my fix for the below in in stead.

https://github.com/dlfcn-win32/dlfcn-win32/issues/112

As for the name of error 0 changing, just doesn't feel right
to use a boolean for a variable that records error codes.
ERROR_SUCCESS didn't feel right either in some cases so I used
what I found below it in my copy of winerr.h which also used
the value 0, NO_ERROR.

Should compile, we'll see